+++ backend/src/strava/__tests__/strava-auth.service.tests.ts
@@ -84,9 +84,10 @@ describe('StravaAuthServiceに関するテスト', () => {
       expect.unreachable('例外が投げられるはず');
     } catch (error) {
       expect(error).toBeInstanceOf(AppException);
-      expect((error as AppException).getResponse()).toEqual(
-        expect.objectContaining({ errorCode: APP_ERROR_CODE.stravaAuthFailed })
-      );
+      if (!(error instanceof AppException)) {
+        throw error;
+      }
+      expect(error.getResponse()).toEqual(expect.objectContaining({ errorCode: APP_ERROR_CODE.stravaAuthFailed }));
     }
   });
 });